+++ docs/changelog.md
@@ -1,4 +1,9 @@
 # TBD
+### Features
+* Lighthouse images are run with `RUST_BACKTRACE=full`
+
+### Changes
+* Upgraded to module-api-lib 0.15.0 for latest Kurtosis compatibility
 
 # 0.4.18
 ### Features